fixes R16Pixel video type

+++ VL.Core/src/Types/Video/Pixels/R16Pixel.cs
@@ -3,7 +3,7 @@ using VL.Lib.Basics.Imaging;
 
 namespace VL.Lib.Basics.Video
 {
-    public record struct R16Pixel(byte R) : IPixel
+    public record struct R16Pixel(ushort R) : IPixel
     {
         public PixelFormat PixelFormat => PixelFormat.R16;
     }
